if MPI_Allreduce uses MPI_UNSIGNED_LONG, the associated variable should be unsigned long! This leads to troubles on 64bit platforms. This error was detected by Dominik Rössler.
git-svn-id: https://scm.projects.hlrs.de/svn/ls1/MarDyn/trunk@50 a63bd714-7e14-4b5e-94e7-302c8c8ff188

+++ src/parallel/DomainDecomposition.cpp
@@ -252,7 +252,7 @@ void parallel::DomainDecomposition::reducevalues(double *Upot, double *Virial, d
   double Virial_glob;
   double summv2_glob;
   double sumIw2_glob;
-  int num_mol_glob;
+  unsigned long num_mol_glob;
   MPI_Allreduce(num_mol, &num_mol_glob, 1, MPI_UNSIGNED_LONG, MPI_SUM, comm_topology);
   MPI_Allreduce(Upot, &Upot_glob, 1, MPI_DOUBLE, MPI_SUM, comm_topology);
   MPI_Allreduce(Virial, &Virial_glob, 1, MPI_DOUBLE, MPI_SUM, comm_topology);
